Changes based on Oscars feedback
