+++ scalding-core/src/main/scala/com/twitter/scalding/ExecutionUtil.scala
@@ -1,70 +1,8 @@
 package com.twitter.scalding
 
-import scala.collection.mutable
-
-import com.twitter.algebird.Monoid
-import scala.concurrent.{ Future, Promise }
-import scala.util.{ Failure, Success }
+import com.twitter.algebird.Semigroup
 
 object ExecutionUtil {
-  private[ExecutionUtil] class AsyncSemaphore(initialPermits: Int = 0) {
-    private[this] val waiters = new mutable.Queue[() => Unit]
-    private[this] var availablePermits = initialPermits
-
-    private[ExecutionUtil] class SemaphorePermit {
-      def release() =
-        AsyncSemaphore.this.synchronized {
-          availablePermits += 1
-          if (availablePermits > 0 && waiters.nonEmpty) {
-            availablePermits -= 1
-            waiters.dequeue()()
-          }
-        }
-    }
-
-    def acquire(): Future[SemaphorePermit] = {
-      val promise = Promise[SemaphorePermit]()
-
-      def setAcquired(): Unit =
-        promise.success(new SemaphorePermit)
-
-      synchronized {
-        if (availablePermits > 0) {
-          availablePermits -= 1
-          setAcquired()
-        } else {
-          waiters.enqueue(setAcquired)
-        }
-      }
-
-      promise.future
-    }
-  }
-
-  /**
-   * Run a sequence of executions but only permitting parallelism amount to run at the
-   * same time.
-   *
-   * @param executions List of executions to run
-   * @param parallelism Number to run in parallel
-   * @return Execution Seq
-   */
-  def withParallelism[T](executions: Seq[Execution[T]], parallelism: Int): Execution[Seq[T]] = {
-    require(parallelism > 0, s"Parallelism must be > 0: $parallelism")
-    val sem = new AsyncSemaphore(parallelism)
-
-    def waitRun(e: Execution[T]): Execution[T] = {
-      Execution.fromFuture(_ => sem.acquire())
-        .flatMap(p => e.liftToTry.map((_, p)))
-        .onComplete {
-          case Success((_, p)) => p.release()
-          case Failure(_) => ()
-        }
-        .map(_._1.get)
-    }
-    Execution.sequence(executions.map(waitRun))
-  }
-
   /**
    * Generate a list of executions from a date range
    *
@@ -87,7 +25,7 @@ object ExecutionUtil {
   def runDatesWithParallelism[T](duration: Duration, parallelism: Int = 1)(fn: DateRange => Execution[T])(implicit dr: DateRange): Execution[Seq[(DateRange, T)]] = {
 
     val dates = dr.each(duration).toSeq
-    withParallelism(dates.map(fn), parallelism).map(e => dates.zip(e))
+    Execution.withParallelism(dates.map(fn), parallelism).map(e => dates.zip(e))
   }
 
   /**
@@ -117,6 +55,10 @@ object ExecutionUtil {
    * algebird supports monoids for maps and hll) and you wanted to do a
    * final aggregation of the Monoids computed for each duration.
    */
-  def runDateRangeWithParallelismSum[T](duration: Duration, parallelism: Int = 1)(fn: DateRange => Execution[T])(implicit dr: DateRange, mon: Monoid[T]): Execution[T] =
-    runDateRangeWithParallelism(duration, parallelism)(fn)(dr).map(Monoid.sum[T])
+  def runDateRangeWithParallelismSum[T](duration: Duration, parallelism: Int = 1)(fn: DateRange => Execution[T])(implicit dr: DateRange, semigroup: Semigroup[T]): Execution[T] = {
+    require(dr.each(duration).nonEmpty, s"Date Range can not be empty")
+
+    runDateRangeWithParallelism(duration, parallelism)(fn)(dr)
+      .map(_.reduceLeft[T]{ case (l, r) => Semigroup.plus(l, r) })
+  }
 }
