+++ scalding-core/src/test/scala/com/twitter/scalding/ExecutionUtilTest.scala
@@ -37,47 +37,5 @@ class ExecutionUtilTest extends WordSpec with Matchers {
       val result = runDateRangeWithParallelismSum(Days(1))(testJob)
       assert(run(result).get == days.map(d => 1).sum)
     }
-
-    "handle failure" in {
-      val days = dateRange.each(Days(1)).toSeq
-      val result = withParallelism(Seq(Execution.failed(new Exception("failed"))), 1)
-
-      assert(run(result).isFailure)
-    }
-
-    "handle an error running parallel" in {
-      val executions = Execution.failed(new Exception("failed")) :: 0.to(10).map(i => Execution.from[Int](i)).toList
-
-      val result = withParallelism(executions, 3)
-
-      assert(run(result).isFailure)
-    }
-
-    "run in parallel" in {
-      val executions = 0.to(10).map(i => Execution.from[Int](i)).toList
-
-      val result = withParallelism(executions, 3)
-
-      assert(run(result).get == 0.to(10).toSeq)
-    }
-
-    "block correctly" in {
-      var seen = 0
-      def updateSeen(idx: Int) {
-        assert(seen === idx)
-        seen += 1
-      }
-
-      val executions = 0.to(10).map{ i =>
-        Execution
-          .from[Int](i)
-          .map{ i => Thread.sleep(10 - i); i }
-          .onComplete(t => updateSeen(t.get))
-      }.toList.reverse
-
-      val result = withParallelism(executions, 1)
-
-      assert(run(result).get == 0.to(10).reverse)
-    }
   }
 }
